# docs/swagger/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "REST API for JAV metadata scraping and file organization",
        "title": "Javinizer API",
        "termsOfService": "https://github.com/javinizer/javinizer-go",
        "contact": {
            "name": "API Support",
            "url": "https://github.com/javinizer/javinizer-go/issues"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/v1/batch/{id}/movies/batch-exclude": {
            "post": {
                "description": "Exclude multiple movies from a batch job in a single request. Best-effort: excludes as many as possible and returns per-result failures.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "web"
                ],
                "summary": "Bulk exclude movies from batch organization",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Job ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Result IDs to exclude",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_javinizer_javinizer-go_internal_api_contracts.BatchExcludeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_javinizer_javinizer-go_internal_api_contracts.BatchExcludeResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_javinizer_javinizer-go_internal_api_contracts.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_javinizer_javinizer-go_internal_api_contracts.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/v1/batch/{id}/results/{resultId}": {
            "patch": {
                "description": "Update a movie's metadata within a batch job's results",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "web"
                ],
                "summary": "Update movie in batch job",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Job ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Result ID",
                        "name": "resultId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated movie data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_javinizer_javinizer-go_internal_api_contracts.UpdateMovieRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_javinizer_javinizer-go_internal_api_contracts.MovieResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_javinizer_javinizer-go_internal_api_contracts.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_javinizer_javinizer-go_internal_api_contracts.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/batch/{id}/results/{resultId}/exclude": {
            "post": {
                "description": "Mark a movie in a batch job as excluded from file organization",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "web"
                ],
                "summary": "Exclude movie from batch organization",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Job ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Result ID",
                        "name": "resultId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_javinizer_javinizer-go_internal_api_contracts.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/batch/{id}/results/{resultId}/poster-crop": {
            "post": {
                "description": "Re-crop a temp poster for the review page using fixed-size crop coordinates",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "web"
                ],
                "summary": "Update manual poster crop in batch job",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Job ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Result ID",
                        "name": "resultId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Crop coordinates",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_javinizer_javinizer-go_internal_api_contracts.PosterCropRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_javinizer_javinizer-go_internal_api_contracts.PosterCropResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_javinizer_javinizer-go_internal_api_contracts.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_javinizer_javinizer-go_internal_api_contracts.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/batch/{id}/results/{resultId}/poster-from-url": {
            "post": {
                "description": "Download a poster image from a URL and set it as the movie's poster in the batch job",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "web"
                ],
                "summary": "Download poster from URL",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Job ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Result ID",
                        "name": "resultId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Poster URL",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_javinizer_javinizer-go_internal_api_contracts.PosterFromURLRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_javinizer_javinizer-go_internal_api_contracts.PosterFromURLResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_javinizer_javinizer-go_internal_api_contracts.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_javinizer_javinizer-go_internal_api_contracts.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_javinizer_javinizer-go_internal_api_contracts.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_javinizer_javinizer-go_internal_api_contracts.BatchExcludeFailed": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "Result not found in job"
                },
                "result_id": {
                    "type": "string",
                    "example": "uuid-1"
                }
            }
        },
        "github_com_javinizer_javinizer-go_internal_api_contracts.BatchExcludeRequest": {
            "type": "object",
            "required": [
                "result_ids"
            ],
            "properties": {
                "result_ids": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "uuid-1",
                        "uuid-2"
                    ]
                }
            }
        },
        "github_com_javinizer_javinizer-go_internal_api_contracts.BatchExcludeResponse": {
            "type": "object",
            "properties": {
                "excluded": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "failed": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_javinizer_javinizer-go_internal_api_contracts.BatchExcludeFailed"
                    }
                },
                "job": {
                    "$ref": "#/definitions/github_com_javinizer_javinizer-go_internal_api_contracts.BatchJobResponse"
                }
            }
        },
...
        "github_com_javinizer_javinizer-go_internal_api_contracts.BatchJobResponse": {
            "type": "object",
            "properties": {
                "completed": {
                    "type": "integer"
                },
                "completed_at": {
                    "type": "string"
                },
                "destination": {
                    "type": "string"
                },
                "excluded": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "boolean"
                    }
                },
                "failed": {
                    "type": "integer"
                },
                "files": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "id": {
                    "type": "string"
                },
                "operation_count": {
                    "type": "integer"
                },
                "operation_mode_override": {
                    "$ref": "#/definitions/github_com_javinizer_javinizer-go_internal_operationmode.OperationMode"
                },
                "persist_error": {
                    "type": "string"
                },
                "progress": {
                    "type": "number"
                },
                "results": {
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/github_com_javinizer_javinizer-go_internal_api_contracts.BatchFileResult"
                    }
                },
                "reverted_count": {
                    "type": "integer"
                },
                "started_at": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/github_com_javinizer_javinizer-go_internal_models.JobStatus"
                },
                "total_files": {
                    "type": "integer"
                },
                "update": {
                    "type": "boolean"
                }
            }
        },
...
        "github_com_javinizer_javinizer-go_internal_api_contracts.DataSource": {
            "type": "object",
            "properties": {
                "confidence": {
                    "description": "Confidence score (0.0-1.0)",
                    "type": "number",
                    "example": 0.9
                },
                "last_updated": {
                    "description": "ISO 8601 timestamp",
                    "type": "string",
                    "example": "2024-01-15T10:30:00Z"
                },
                "source": {
                    "description": "\"scraper\" or \"nfo\"",
                    "type": "string",
                    "example": "nfo"
                }
            }
        },
        "github_com_javinizer_javinizer-go_internal_api_contracts.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "Movie not found"
                },
                "errors": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
        "github_com_javinizer_javinizer-go_internal_api_contracts.MergeStatistics": {
            "type": "object",
            "properties": {
                "conflicts_resolved": {
                    "type": "integer",
                    "example": 5
                },
                "empty_fields": {
                    "type": "integer",
                    "example": 2
                },
                "from_nfo": {
                    "type": "integer",
                    "example": 3
                },
                "from_scraper": {
                    "type": "integer",
                    "example": 10
                },
                "merged_arrays": {
                    "type": "integer",
                    "example": 2
                },
                "total_fields": {
                    "type": "integer",
                    "example": 15
                }
            }
        },
        "github_com_javinizer_javinizer-go_internal_api_contracts.MovieResponse": {
            "type": "object",
            "properties": {
                "merge_stats": {
                    "description": "Merge statistics when NFO merging occurred",
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_javinizer_javinizer-go_internal_api_contracts.MergeStatistics"
                        }
                    ]
                },
                "movie": {
                    "$ref": "#/definitions/github_com_javinizer_javinizer-go_internal_api_contracts.MovieView"
                },
                "provenance": {
                    "description": "Field-level data source tracking",
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/github_com_javinizer_javinizer-go_internal_api_contracts.DataSource"
                    }
                }
            }
        },
...
        "github_com_javinizer_javinizer-go_internal_api_contracts.PosterCropRequest": {
            "type": "object",
            "properties": {
                "height": {
                    "type": "integer",
                    "minimum": 1
                },
                "max_poster_height": {
                    "description": "MaxPosterHeight optional override for the max poster height (px). 0 = no cap.\nWhen omitted, the configured output.max_poster_height is used.",
                    "type": "integer"
                },
                "width": {
                    "type": "integer",
                    "minimum": 1
                },
                "x": {
                    "type": "integer",
                    "minimum": 0
                },
                "y": {
                    "type": "integer",
                    "minimum": 0
                }
            }
        },
        "github_com_javinizer_javinizer-go_internal_api_contracts.PosterCropResponse": {
            "type": "object",
            "properties": {
                "cropped_poster_url": {
                    "type": "string"
                }
            }
        },
        "github_com_javinizer_javinizer-go_internal_api_contracts.PosterFromURLRequest": {
            "type": "object",
            "required": [
                "url"
            ],
            "properties": {
                "url": {
                    "type": "string"
                }
            }
        },
        "github_com_javinizer_javinizer-go_internal_api_contracts.PosterFromURLResponse": {
            "type": "object",
            "properties": {
                "cropped_poster_url": {
                    "type": "string"
                },
                "poster_url": {
                    "type": "string"
                }
            }
        },
...
        "github_com_javinizer_javinizer-go_internal_api_contracts.UpdateMovieRequest": {
            "type": "object",
            "required": [
                "movie"
            ],
            "properties": {
                "movie": {
                    "$ref": "#/definitions/github_com_javinizer_javinizer-go_internal_api_contracts.MovieView"
                }
            }
        },
...
        "github_com_javinizer_javinizer-go_internal_operationmode.OperationMode": {
            "type": "string",
            "enum": [
                "organize",
                "in-place",
                "in-place-norenamefolder",
                "metadata-artwork",
                "preview"
            ],
            "x-enum-varnames": [
                "OperationModeOrganize",
                "OperationModeInPlace",
                "OperationModeInPlaceNoRenameFolder",
                "OperationModeMetadataArtwork",
                "OperationModePreview"
            ]
        },
...
    }
}